+++ src/core/client/startup.ts
@@ -19,6 +19,7 @@ import './streamers/item';
 import './streamers/marker';
 import './streamers/object';
 import './streamers/ped';
+import './streamers/polygon';
 import './streamers/textlabel';
 import './streamers/worldNotifications';